chore: move tests to separate file
